Cartoonify/requirements.txt
opencv-python
easygui
numpy
imageio
matplotlib
tk
pillow